backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
python-docx==1.1.0
sentence-transformers==2.5.1
faiss-cpu
groq>=0.11.0
python-dotenv==1.0.1
numpy==1.26.4
pymupdf==1.24.0
httpx<0.28